GameObject.ObjectID`: Change to ushort * It was an int to allow -1 to work as an invalid ID, but this was redundant since `Region` can't give 0 as ID. * It also required casting to ushort pretty much everywhere, since object IDs are sent as ushort in most packets. * It also introduced a subtle bug where casting an invalid id (-1) to ushort resulted in a valid ID (65535). This happened in `LosCheckHandler` and potentially other places.

+++ GameServer/gameutils/LosCheckHandler.cs
@@ -49,8 +49,8 @@ namespace DOL.GS
 
         public bool StartLosCheck(GameObject source, GameObject target, CheckLosResponse callback)
         {
-            ushort sourceObjectId = (ushort) source.ObjectID;
-            ushort targetObjectId = (ushort) target.ObjectID;
+            ushort sourceObjectId = source.ObjectID;
+            ushort targetObjectId = target.ObjectID;
 
             if (sourceObjectId == 0 || targetObjectId == 0)
                 return false;
